% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/swt.R
\name{d2_temp_lifeport}
\alias{d2_temp_lifeport}
\title{Calculate Mahalanobis distance D-square for LifePort temperature data.}
\usage{
d2_temp_lifeport(data)
}
\arguments{
\item{data}{data frame or matrix with temperature data}
}
\value{
vector with D-square for temperature
}
\description{
Calculate Mahalanobis distance D-square for LifePort temperature data.
}
