% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/swt.R
\name{d2prc_temp_lifeport}
\alias{d2prc_temp_lifeport}
\title{Returns the percentile ranke of the distance D-squared for the temperature}
\usage{
d2prc_temp_lifeport(d2)
}
\arguments{
\item{d2}{D-squared}
}
\value{
percentile rank
}
\description{
Returns the percentile ranke of the distance D-squared for the temperature
}
