% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/swt.R
\name{date2num}
\alias{date2num}
\title{Convert POSIXct data type (date/time) to Excel days since origin}
\usage{
date2num(dates)
}
\arguments{
\item{dates}{character string in the form of YYYY-mm-dd}
}
\value{
number of days
}
\description{
Convert POSIXct data type (date/time) to Excel days since origin
}
