% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/swt.R
\name{egfr_ckd_epi}
\alias{egfr_ckd_epi}
\title{CKD-EPI Creatinine Equation (2021)}
\usage{
egfr_ckd_epi(SCr, age, sex, units = "SI")
}
\arguments{
\item{SCr}{serum creatinine in mg/dL (US) or umol/L (S)}

\item{age}{age in years}

\item{sex}{either "F" for female, or "M" for male}

\item{units}{unit for SCr, either "SI" (umol/L; default) or "US" (mg/dL)}
}
\value{
eGFR in mL/min/1.73m2
}
\description{
Calculates eGFR according to the 2021 formula.
}
\details{
See equation and references at \url{https://www.kidney.org/ckd-epi-creatinine-equation-2021}.
}
