% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/swt.R
\name{egfr_schwartz}
\alias{egfr_schwartz}
\title{Revised Schwartz Equation for eGFR (2009)
see https://www.mdcalc.com/calc/10008/revised-schwartz-equation-glomerular-filtration-rate-gfr-2009#evidence}
\usage{
egfr_schwartz(SCr, height, units = "SI")
}
\arguments{
\item{SCr}{Serum creatinine in mg/dL (US) or μmol/L (S)}

\item{height}{in cm}

\item{units}{unit for SCr, either "SI" (μmol/L; default) or "US" (mg/dL)}
}
\value{
eGFR mL/min/1.73m2
}
\description{
Revised Schwartz Equation for eGFR (2009)
see https://www.mdcalc.com/calc/10008/revised-schwartz-equation-glomerular-filtration-rate-gfr-2009#evidence
}
