% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{exam.device}
\alias{exam.device}
\title{EXAM device data}
\format{
A data frame with self explaining variable names:
\describe{
  \item{SerialNumber}{serial number of the device}
  \item{Type}{type note}
  \item{SubType}{subtype note}
  \item{UnitID}{name given to the device}
  \item{FirmwareVersion}{firmware version}
  \item{FileID}{file id}
  \item{StartTime}{start date and time of machine}
  \item{DataState}{data state, if it is complete}
  \item{HasGaps}{whether data has haps}
  \item{Runtime}{run time of the machine}
  \item{StopTime}{time and date of machine stop}
  \item{Filename}{file name}#'
}
}
\source{
\url{https://data.swisstransplant.org/}
}
\usage{
exam.device
}
\description{
Example data form the LifePort kidney transporter.
}
\keyword{datasets}
