% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{exam.organ}
\alias{exam.organ}
\title{EXAM organ data}
\format{
A data frame with self explaining variable names:
\describe{
  \item{OrganID}{identifier entered into the machine}
  \item{KidneySide}{left or right kidney}
  \item{BloodType}{self explaining}
  \item{CrossClampTime.Date}{self explaining}
  \item{CrossClampTimezone}{self explaining}
  \item{TotalIschemicTime}{self explaining}
  \item{PerfusateLot}{self explaining}
  \item{PerfusateExpirationDate}{self explaining}
  \item{PerfusateUsed}{self explaining}
  \item{Cannula}{self explaining}
  \item{CannulaExpirationDate}{self explaining}
  \item{CassetteLot.}{self explaining}
  \item{CasetteExpirationDate}{self explaining}
  \item{ID}{self explaining}
  \item{DonorID}{self explaining}
}
}
\source{
\url{https://data.swisstransplant.org/}
}
\usage{
exam.organ
}
\description{
Example data form the LifePort kidney transporter.
}
\keyword{datasets}
