% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{exam.timeseries}
\alias{exam.timeseries}
\title{EXAM time series data}
\format{
A list of data frames with time series data:
\describe{
  \item{SerialNumber}{serial number of the device}
  \item{FlowRate }{flow rate}
  ...
}
}
\source{
\url{https://data.swisstransplant.org/}
}
\usage{
exam.timeseries
}
\description{
Example data form the LifePort kidney transporter.
}
\keyword{datasets}
