% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/swt.R
\name{finalise_plot}
\alias{finalise_plot}
\title{Arrange alignment and save SWT plot}
\usage{
finalise_plot(
  plot_name,
  source_name,
  save_filepath,
  width_pixels = 640,
  height_pixels = 450,
  logo_image_path = file.path(system.file(c("inst", "extdata"), package = "swt"),
    "swtlogo.png"),
  footer = TRUE
)
}
\arguments{
\item{plot_name}{The ggplot object of the plot}

\item{source_name}{The text in the footer}

\item{save_filepath}{Exact filepath that you want the plot to be saved to}

\item{width_pixels}{Width in pixels - defaults to 640}

\item{height_pixels}{Height in pixels - defaults to 450}

\item{logo_image_path}{File path for the logo image}

\item{footer}{Whether to use a footer with the logo (TRUE or FALSE)}
}
\value{
(Invisibly) an updated ggplot object.
}
\description{
Running this function will save your plot with the correct guidelines for
publication for a SWT graphic. It will left align the title and subtitle,
add the SWT logo at the bottom right and save it to your specified location.
}
