% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/swt.R
\name{fmt_hla}
\alias{fmt_hla}
\title{Format HLA}
\usage{
fmt_hla(v_char)
}
\arguments{
\item{v_char}{character vector}
}
\value{
formatted character vector
}
\description{
Helper function to format strings for broads, e.g. A(10) becomes A10 and A becomes NA.
}
