% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/swt.R
\name{freq_perc}
\alias{freq_perc}
\title{Returns frequency count and percentage}
\usage{
freq_perc(x, count.na = TRUE, d2 = 1)
}
\arguments{
\item{x}{logical vector}

\item{count.na}{count NAs in denominator}

\item{d2}{number of digits}
}
\value{
character object
}
\description{
Helper function for tidy formatting.
}
