% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/swt.R
\name{get_days_in_year}
\alias{get_days_in_year}
\title{Get the number of days in a year. Used in survival analysis to convert event times.}
\usage{
get_days_in_year()
}
\value{
number of days
}
\description{
Get the number of days in a year. Used in survival analysis to convert event times.
}
