% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/swt.R
\name{HLA_mismatch}
\alias{HLA_mismatch}
\title{The function calculates HLA mismatches.}
\usage{
HLA_mismatch(
  D.A1,
  D.A2,
  D.B1,
  D.B2,
  D.DR1,
  D.DR2,
  R.A1,
  R.A2,
  R.B1,
  R.B2,
  R.DR1,
  R.DR2
)
}
\arguments{
\item{D.A1}{Donor HLA Antigen on allele 1 locus A}

\item{D.A2}{Donor HLA Antigen on allele 2 locus A}

\item{D.B1}{Donor HLA Antigen on allele 1 locus B}

\item{D.B2}{Donor HLA Antigen on allele 2 locus B}

\item{D.DR1}{Donor HLA Antigen on allele 1 locus DR}

\item{D.DR2}{Donor HLA Antigen on allele 2 locus DR}

\item{R.A1}{Recipient HLA Antigen on allele 1 locus A}

\item{R.A2}{Recipient HLA Antigen on allele 2 locus A}

\item{R.B1}{Recipient HLA Antigen on allele 1 locus B}

\item{R.B2}{Recipient HLA Antigen on allele 2 locus B}

\item{R.DR1}{Recipient HLA Antigen on allele 1 locus DR}

\item{R.DR2}{Recipient HLA Antigen on allele 2 locus DR}
}
\value{
data frame with mismatch information.
}
\description{
The function calculates HLA mismatches.
}
