% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/swt.R
\name{hla_parse}
\alias{hla_parse}
\title{Parser for the unstructured SOAS HLA information into structured data.}
\usage{
hla_parse(D_HLA, R_HLA)
}
\arguments{
\item{D_HLA}{Donor HLA antigens. Character string from SOAS variable D HLA Ag.}

\item{R_HLA}{Recipient HLA antigens. Character string from SOAS variable R HLA Ag.}
}
\value{
a data frame with structured HLA information.
}
\description{
Parser for the unstructured SOAS HLA information into structured data.
}
