% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/swt.R
\name{kidmo}
\alias{kidmo}
\title{KIDMO Score}
\usage{
kidmo(
  D_age = 55,
  D_deathcause = "cerebral hemorrhage",
  D_diabetes = FALSE,
  D_hypertension = FALSE,
  R_age = 57,
  R_retpx = FALSE,
  R_tpxyear = 2026,
  times = c(2, 5)
)
}
\arguments{
\item{D_age}{donor age in years}

\item{D_deathcause}{donor cause of death (cerebral hemorrhage, anoxia, or others)}

\item{D_diabetes}{donor history of diabetes (binary)}

\item{D_hypertension}{donor history of hypertension (binary)}

\item{R_age}{recipient age in years}

\item{R_retpx}{recipient listed for retransplant (binary)}

\item{R_tpxyear}{recipient year of transplant (continuous)}

\item{times}{time points for predictions, in years}
}
\value{
KIDMO Score
}
\description{
Calculates the KIDMO Score.
}
