% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/swt.R
\name{kidmo_ecdf}
\alias{kidmo_ecdf}
\title{KIDMO empirical cumulative distribution function (ecdf)}
\usage{
kidmo_ecdf(log_hr)
}
\arguments{
\item{log_hr}{log hazard ratio (unscaled)}
}
\value{
percentile rank
}
\description{
Conversion of log hazard ratio into percentile rank.
}
