% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/swt.R
\name{kidmo_hr2rank}
\alias{kidmo_hr2rank}
\title{KIDMO conversion of hazard ratio to percentile rank.}
\usage{
kidmo_hr2rank(hr)
}
\arguments{
\item{hr}{hazard ratio}
}
\value{
percentile
}
\description{
KIDMO conversion of hazard ratio to percentile rank.
}
