% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/swt.R
\name{lifeport_d2}
\alias{lifeport_d2}
\title{D-squared for LifePort data}
\usage{
lifeport_d2(data, type)
}
\arguments{
\item{data}{data frame or matrix with temperature or perfusion data}

\item{type}{string, type of D-square either "temp" or "perf"}
}
\value{
vector with D-squared
}
\description{
Calculate Mahalanobis distance D-squared for LifePort temperature and perfusion data.
}
