% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/swt.R
\name{lifeport_d2_temp}
\alias{lifeport_d2_temp}
\title{Calculate Mahalanobis distance D-square for LifePort temperature data.}
\usage{
lifeport_d2_temp(data)
}
\arguments{
\item{data}{data frame or matrix with temperature data}
}
\value{
vector with D-square for temperature
}
\description{
Calculate Mahalanobis distance D-square for LifePort temperature data.
}
