% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/swt.R
\name{lifeport_d2prc_temp}
\alias{lifeport_d2prc_temp}
\title{Returns the percentile rank of the distance D-squared for the temperature.}
\usage{
lifeport_d2prc_temp(d2)
}
\arguments{
\item{d2}{D-squared}
}
\value{
percentile rank
}
\description{
Returns the percentile rank of the distance D-squared for the temperature.
}
