% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/swt.R
\name{lifeport_d2toRank}
\alias{lifeport_d2toRank}
\title{Returns the percentile rank of the temperature or perfusion D-squared.}
\usage{
lifeport_d2toRank(d2, type)
}
\arguments{
\item{d2}{D-squared}

\item{type}{string, type of D-square either "temp" or "perf"}
}
\value{
percentile rank
}
\description{
Returns the percentile rank of the temperature or perfusion D-squared.
}
