% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/swt.R
\name{lifeport_read}
\alias{lifeport_read}
\title{Read LifePort data}
\usage{
lifeport_read(file, format = "guess")
}
\arguments{
\item{file}{The data file}

\item{format}{guess, binary or plaintxt (default guess)}
}
\value{
a list with LifePort data
}
\description{
Read LifePort data
}
