% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/swt.R
\name{median_iqr}
\alias{median_iqr}
\title{Returns median and interquartile range IQR}
\usage{
median_iqr(x, d1 = 1, d2 = 1, d3 = 1, compact = FALSE)
}
\arguments{
\item{x}{numeric vector}

\item{d1}{number of digits}

\item{d2}{number of digits}

\item{d3}{number of digits}

\item{compact}{use en dash instead of "from X to Y"}
}
\value{
character object
}
\description{
Helper function for tidy formatting.
}
