% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/swt.R
\name{miss_perc}
\alias{miss_perc}
\title{Returns frequency count and percentage of missing data.}
\usage{
miss_perc(x, d2 = 1)
}
\arguments{
\item{x}{vector}

\item{d2}{number of digits}
}
\value{
character object
}
\description{
Helper function for tidy formatting.
}
