% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/swt.R
\name{nearest}
\alias{nearest}
\title{Nearest element}
\usage{
nearest(y, q)
}
\arguments{
\item{y}{vector to be searched}

\item{q}{vector of values of interest}
}
\value{
indices of the nearest elements in y for a set of values in q
}
\description{
Nearest element in vector for a given set of values.
}
