% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/swt.R
\name{num2date}
\alias{num2date}
\title{Convert Excel days since origin to POSIXct data type (date/time)}
\usage{
num2date(
  days,
  origin = "1899-12-30",
  tz = "CET",
  filter = TRUE,
  pattern = "[0-9]{2}\\\\.[0-9]{2}\\\\.[0-9]{4}",
  format = "\%d.\%m.\%Y",
  round = TRUE
)
}
\arguments{
\item{days}{days since origin as numeric or string}

\item{origin}{origin, default in excel is 1899-12-30}

\item{tz}{time zone to be forced upon}

\item{filter}{a fix for dates not recognized (default is TRUE)}

\item{pattern}{the pattern to find dates not recognized}

\item{format}{format to convert dates not recognized, e.g. \%d.\%m.\%Y \%H:\%M:\%OS}

\item{round}{recommended when format has no time, only date information}
}
\value{
date of the type POSIXct
}
\description{
Convert Excel days since origin to POSIXct data type (date/time)
}
