% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/swt.R
\name{optn_kdri}
\alias{optn_kdri}
\title{OPTN KDRI}
\usage{
optn_kdri(
  D_age,
  D_height,
  D_weight,
  D_hypertension,
  D_diabetes,
  D_CVA,
  D_SCr,
  D_DCD,
  scaling = 1.40436817065005
)
}
\arguments{
\item{D_age}{donor age in years}

\item{D_height}{donor height in cm}

\item{D_weight}{donor weight in kg}

\item{D_hypertension}{donor hypertension}

\item{D_diabetes}{donor diabetes}

\item{D_CVA}{donor cause of death is cardiovascular accident}

\item{D_SCr}{serum creatinine in mg/dL}

\item{D_DCD}{donation after cardiac death}

\item{scaling}{scaling factor that is published every year by the OPTN}
}
\value{
KDRI hazard ratio
}
\description{
OPTN KDRI
}
