% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/swt.R
\name{print.kidmo}
\alias{print.kidmo}
\title{Print method for kidmo objects}
\usage{
\method{print}{kidmo}(x, ...)
}
\arguments{
\item{x}{An object of class \code{kidmo}.}

\item{...}{Additional arguments (ignored).}
}
\description{
Print method for kidmo objects
}
