% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/swt.R
\name{process.lifeport}
\alias{process.lifeport}
\title{Process LifePort data. Adds runtime, clock time vectors, and filtered time
series.}
\usage{
process.lifeport(lpdat, window_size = 30)
}
\arguments{
\item{lpdat}{A list with data from read.lifeport()}

\item{window_size}{rolling window size for filtering}
}
\value{
a list with additional processed data tables
}
\description{
Process LifePort data. Adds runtime, clock time vectors, and filtered time
series.
}
