% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/swt.R
\name{read.lifeport}
\alias{read.lifeport}
\title{Read LifePort data}
\usage{
read.lifeport(file, binary = FALSE)
}
\arguments{
\item{file}{The data file}

\item{binary}{Whether the data file is binary (default FALSE)}
}
\value{
a list with LifePort data
}
\description{
Read LifePort data
}
