% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/swt.R
\name{sumstats_lifeport}
\alias{sumstats_lifeport}
\title{Summary statistics  for LifePort data.}
\usage{
sumstats_lifeport(lpdat)
}
\arguments{
\item{lpdat}{A list with data from read.lifeport()}
}
\value{
a list with additional summary statistics
}
\description{
Summary statistics  for LifePort data.
}
