% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/swt.R
\name{swt_colors}
\alias{swt_colors}
\title{SWT colors}
\usage{
swt_colors()
}
\value{
a SWT color object
}
\description{
Easy access to official SWT color scheme
}
\examples{
mycolors = swt_colors()
mycolors$red.liver

}
