% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/swt.R
\name{swt_document_skeleton}
\alias{swt_document_skeleton}
\title{SWT skeleton for documents}
\usage{
swt_document_skeleton(path)
}
\arguments{
\item{path}{project path}
}
\description{
This internal function supports Swisstransplant Documents in Quarto for RStudio projects.
}
