% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/swt.R
\name{swt_LifePortCaseReport}
\alias{swt_LifePortCaseReport}
\title{Create SWT LifePort Case Report in MS Word}
\usage{
swt_LifePortCaseReport(data.file, output.file, template.file)
}
\arguments{
\item{data.file}{Lifeport data file}

\item{output.file}{target file docx}

\item{template.file}{template file docx}
}
\description{
Create SWT LifePort Case Report in MS Word
}
