% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/swt.R
\name{swt_skeleton}
\alias{swt_skeleton}
\title{SWT skeleton}
\usage{
swt_skeleton(path)
}
\arguments{
\item{path}{project path}
}
\description{
This internal function enables a Swisstransplant Document in Quarto for RStudio projects.
}
