% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/swt.R
\name{swt_slides_skeleton}
\alias{swt_slides_skeleton}
\title{SWT skeleton for slides}
\usage{
swt_slides_skeleton(path)
}
\arguments{
\item{path}{project path}
}
\description{
This internal function supports Swisstransplant Presentations in Quarto for RStudio projects.
}
