% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/swt.R
\name{swt_style}
\alias{swt_style}
\title{SWT theme for ggplot}
\usage{
swt_style(
  title_size = 18,
  subtitle_size = 14,
  font_size = 14,
  grey_theme = FALSE,
  legend_position = "top"
)
}
\arguments{
\item{title_size}{The font size of the title}

\item{subtitle_size}{The font size of the subtitle}

\item{font_size}{The font font size of the legend, axis text, and axis titles}

\item{grey_theme}{Whether to use the grey theme instead (TRUE or FALSE)}

\item{legend_position}{Position of the legend (top, bottom, left or right)}
}
\description{
This function allows you to add the SWT theme to your ggplot graphics.
}
\examples{
\donttest{
library(ggplot2)
ggplot(mtcars, aes(wt, mpg)) +
   geom_point() +
   swt_style()
}

}
