% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/swt.R
\name{tidy_pvalues}
\alias{tidy_pvalues}
\title{Formats p-values.}
\usage{
tidy_pvalues(x, compact = FALSE)
}
\arguments{
\item{x}{numerical vector with p-values}

\item{compact}{logical, no asterisks when TRUE}
}
\value{
formatted p-values as character vector
}
\description{
Helper function for tidy formatting.
}
