% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/swt.R
\name{UK_DCD_Score}
\alias{UK_DCD_Score}
\title{UK DCD Risk Score by Schlegel et al.}
\usage{
UK_DCD_Score(D_age, D_BMI, fWIT, CIT, R_age, R_MELD, retpx)
}
\arguments{
\item{D_age}{donor age in years}

\item{D_BMI}{donor BMI in kg/m^2}

\item{fWIT}{functional warm ischemia time in minutes}

\item{CIT}{cold ischemia time in hours}

\item{R_age}{recipient age in years}

\item{R_MELD}{recipient lab MELD score}

\item{retpx}{whether the aim is a retransplant}
}
\value{
UK DCD Risk Score
}
\description{
UK DCD Risk Score by Schlegel et al.
}
