% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/swt.R
\name{uk_kdri}
\alias{uk_kdri}
\title{UK KDRI 2019}
\usage{
uk_kdri(D_age, D_height, D_hypertension, D_female, D_CMV, D_eGFR, D_days_hosp)
}
\arguments{
\item{D_age}{donor age in years}

\item{D_height}{donor height in cm}

\item{D_hypertension}{donor hypertension}

\item{D_female}{donor is female}

\item{D_CMV}{donor cytomegalovirus positive}

\item{D_eGFR}{estimated glomerular filtration rate (eGFR) in mL/min/1.73m2}

\item{D_days_hosp}{days in hospital}
}
\value{
UK KDRI 2019 hazard ratio
}
\description{
Calculates the UK KDRI version from 2019.
}
\details{
Reference: Kim JJ, Curtis RMK, Reynolds B, et al. The UK kidney donor risk index poorly predicts long-term transplant survival in paediatric kidney transplant recipients. Front Immunol. 2023;14:1207145. doi:10.3389/fimmu.2023.1207145

Calculator at \url{https://www.glasgowtransplant.com/tools/ukkdri.html}.
}
