% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/swt.R
\name{egfr_ckd_epi}
\alias{egfr_ckd_epi}
\title{CKD-EPI Creatinine Equation for eGFR (2021)
see https://www.kidney.org/content/ckd-epi-creatinine-equation-2021}
\usage{
egfr_ckd_epi(SCr, age, sex, units = "SI")
}
\arguments{
\item{SCr}{Serum creatinine in mg/dL (US) or μmol/L (S)}

\item{age}{age in years}

\item{sex}{either "F" for female, or "M" for male}

\item{units}{unit for SCr, either "SI" (μmol/L; default) or "US" (mg/dL)}
}
\value{
eGFR mL/min/1.73m2
}
\description{
CKD-EPI Creatinine Equation for eGFR (2021)
see https://www.kidney.org/content/ckd-epi-creatinine-equation-2021
}
