% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/swt.R
\name{egfr_schwartz}
\alias{egfr_schwartz}
\title{Revised Schwartz Equation (2009)}
\usage{
egfr_schwartz(SCr, height, units = "SI")
}
\arguments{
\item{SCr}{serum creatinine in mg/dL (US) or umol/L (S)}

\item{height}{height in cm}

\item{units}{unit for SCr, either "SI" (umol/L; default) or "US" (mg/dL)}
}
\value{
eGFR in mL/min/1.73m2
}
\description{
Calculates eGFR for pediatric patients.
}
\details{
See equation and examples at \url{https://www.mdcalc.com/calc/10008/revised-schwartz-equation-glomerular-filtration-rate-gfr-2009#evidence}.
}
