% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{exam.device}
\alias{exam.device}
\title{EXAM device data}
\format{
## `exam.device`
A data frame with self explaining variable names:
\describe{
  \item{SerialNumber}{serial number of the device}
  \item{UnitID}{name given to the device}
  ...
}
}
\source{
<https://data.swisstransplant.org/>
}
\usage{
exam.device
}
\description{
Example data form the LifePort kidney transporter.
}
\keyword{datasets}
