% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{exam.organ}
\alias{exam.organ}
\title{EXAM organ data}
\format{
## `exam.organ`
A data frame with self explaining variable names:
\describe{
  \item{OrganID}{identifier entered into the machine}
  \item{KidneySide}{left or right kidney}
  ...
}
}
\source{
<https://swtdata.org/>
}
\usage{
exam.organ
}
\description{
Example data form the LifePort kidney transporter.
}
\keyword{datasets}
