% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{exam.sumstats}
\alias{exam.sumstats}
\title{EXAM summary statistics}
\format{
## `exam.sumstats`
A data frame with self explaining variable names:
\describe{
  \item{perfusion.dur}{perfusion duration in minutes}
  \item{perfusion.dur.str}{perfusion duration in HH:MM:SS}
  ...
}
}
\source{
<https://data.swisstransplant.org/>
}
\usage{
exam.sumstats
}
\description{
Example data form the LifePort kidney transporter.
}
\keyword{datasets}
