% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/swt.R
\name{hla_parse}
\alias{hla_parse}
\title{Parse HLA data}
\usage{
hla_parse(D_HLA, R_HLA)
}
\arguments{
\item{D_HLA}{donor HLA antigens; character string from SOAS variable D HLA Ag.}

\item{R_HLA}{recipient HLA antigens; character string from SOAS variable R HLA Ag.}
}
\value{
data frame with structured HLA information
}
\description{
Parser to convert unstructured SOAS HLA information into structured data.
}
