% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/swt.R
\name{lifeport_process}
\alias{lifeport_process}
\title{Process LifePort data. Adds runtime, clock time, and smoothed time series.}
\usage{
lifeport_process(lpdat, window_size = 15)
}
\arguments{
\item{lpdat}{a list with data from lifeport_read()}

\item{window_size}{rolling window size for filtering}
}
\value{
a list with LifePort data
}
\description{
Process LifePort data. Adds runtime, clock time, and smoothed time series.
}
