% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/swt.R
\name{lifeport_read}
\alias{lifeport_read}
\title{Read LifePort raw data}
\usage{
lifeport_read(file, format = "guess")
}
\arguments{
\item{file}{data file with path}

\item{format}{guess (default), binary or plaintxt}
}
\value{
list with LifePort data
}
\description{
Function to read LifePort binary as well as ASCII raw data files.
}
