% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/swt.R
\name{mean_sd}
\alias{mean_sd}
\title{Returns mean and SD as string}
\usage{
mean_sd(x, d1 = 1, d2 = 1)
}
\arguments{
\item{x}{numeric vector}

\item{d1}{number of digits}

\item{d2}{number of digits}
}
\value{
character object
}
\description{
Helper function for tidy formatting.
}
