% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/swt.R
\name{num2date}
\alias{num2date}
\title{Convert excel days since origin to POSIXct data type (date/time)}
\usage{
num2date(days, origin = "1899-12-30", tz = "CET")
}
\arguments{
\item{days}{days since origin as numeric or string}

\item{origin}{origin, default in excel is 1899-12-30}

\item{tz}{time zone to be forced upon}
}
\value{
date of the type POSIXct
}
\description{
Convert excel days since origin to POSIXct data type (date/time)
}
