% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/swt.R
\name{tidy_rmsfit}
\alias{tidy_rmsfit}
\title{Tidy rms model fit results}
\usage{
tidy_rmsfit(fit, ...)
}
\arguments{
\item{fit}{model fit from rms}

\item{...}{optional arguments to summary of the rms fit object.}
}
\value{
formatted data.frame
}
\description{
Shows tidy regression table with results as data frame.
}
