% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/swt.R
\name{uk_dcd_score}
\alias{uk_dcd_score}
\title{UK DCD Risk Score}
\usage{
uk_dcd_score(D_age, D_BMI, fWIT, CIT, R_age, R_MELD, retpx)
}
\arguments{
\item{D_age}{donor age in years}

\item{D_BMI}{donor BMI in kg/m^2}

\item{fWIT}{functional warm ischemia time in minutes}

\item{CIT}{cold ischemia time in hours}

\item{R_age}{recipient age in years}

\item{R_MELD}{recipient lab MELD score}

\item{retpx}{whether the aim is a retransplant}
}
\value{
UK DCD Risk Score
}
\description{
Calculates the UK DCD Risk Score that can range between 0 and 27.
}
\details{
Reference: Schlegel A, Kalisvaart M, Scalera I, et al. The UK DCD Risk Score: A new proposal to define futility in donation-after-circulatory-death liver transplantation. J Hepatol. 2018;68(3):456-464. doi:10.1016/j.jhep.2017.10.034
}
